tested and working with escape sequences
Program outputs escape sequences to erase from
end of line to the right side of screen, and
from the end of the output to the bottom of the
screen.  Escape sequences are output on stderr,
while program output is written to stdout, so you
can redirect output and not get blurred by escapes.

Signed-off-by: Luis Colorado <luiscoloradourcola@gmail.com>